requirements.txt
Django==1.11.5
django-debug-toolbar==1.4
humanize==0.5.1
sqlparse==0.1.19
celery==3.1.25
django-material==0.8.0
pytz
django-mysql
mysqlclient
gunicorn
openpyxl==2.3.2
django-celery==3.1.17
django-bulk-update==2.1.0
djangosaml2
pendulum==1.2.4
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-env==0.6.0
pytest-pep8==1.0.6
mixer==5.6.6
mock==2.0.0
django-widget-tweaks
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
